pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
langchain-core = "^0.2.10"
ruff = "^0.5.0"
langchain-text-splitters = "^0.2.2"
openai = "^1.35.7"
ipykernel = "^6.29.4"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
faiss-cpu = "^1.8.0.post1"
chainlit = "^1.1.305"
streamlit = "^1.36.0"
llama-index = "^0.10.50"
llama-index-llms-openai = "^0.1.24"
llama-index-embeddings-openai = "^0.1.10"
llama-index-llms-azure-openai = "^0.1.8"
llama-index-graph-stores-nebula = "^0.2.0"
llama-index-graph-stores-neo4j = "^0.2.6"
llama-index-embeddings-azure-openai = "^0.1.10"
nest-asyncio = "^1.6.0"
scikit-learn = "^1.5.0"
yfinance = "^0.2.40"
plotly = "^5.22.0"
langchain-experimental = "^0.0.62"
langchainhub = "^0.1.20"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]